CrewAI-Researcher/requirements.txt
crewai
redis
python-dotenv
langchain
openai
anthropic
langchain-community
langchain-anthropic
pydantic